requirements.txt
google-generativeai>=0.3.0
rich>=13.0.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-socketio>=5.0.0
requests>=2.25.0 